requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-cors-headers==3.7.0
django-rest-authtoken==2.1.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==20.9
pycryptodome==3.10.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
